Added info for the most used implementations
Outlined the basics of how to use as a webfont or the raw SVGs. I've used HTML (taken from the Bootstrap docs) so not entirely sure how it will render when live. 

Still need to create the links to each frameworks documentation. I was thinking of having a material design card/tile style interface with each framework having its own card/tile with the icon centered.